AB#32299 split OpenAI service responsibilities
