+++ applications/Unity.GrantManager/modules/Unity.AI/src/Unity.AI.Application/AI/Runtime/IOpenAIConfigurationResolver.cs
@@ -1,0 +1,16 @@
+namespace Unity.AI.Runtime;
+
+public interface IOpenAIConfigurationResolver
+{
+    string ResolveProviderName(string? operationName = null);
+
+    string ResolveApiKey(string? operationName = null);
+
+    string ResolveMaxTokensParameterNameForOperation(string? operationName = null);
+
+    double? ResolveConfiguredTemperature(string? operationName = null);
+
+    int ResolveCompletionTokens(string operationName, int defaultValue);
+
+    string ResolveApiUrl(string? operationName = null);
+}
