+++ applications/Unity.GrantManager/modules/Unity.AI/src/Unity.AI.Application/AI/Runtime/OpenAITransportService.cs
@@ -1,8 +1,6 @@
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.Net;
 using System.Net.Http;
 using System.Text;
@@ -15,17 +13,11 @@ namespace Unity.AI.Runtime;
 
 internal class OpenAITransportService(
     HttpClient httpClient,
-    IConfiguration configuration,
+    IOpenAIConfigurationResolver configurationResolver,
     ILogger<OpenAITransportService> logger) : IOpenAITransportService, ITransientDependency
 {
-    private const string DefaultMaxTokensParameterName = "max_completion_tokens";
-    private const string LegacyMaxTokensParameterName = "max_tokens";
-    private const string DefaultProviderName = "OpenAI";
-    private const string OpenAiApiKeyEnvironmentVariableName = "AZURE_OPENAI_API_KEY";
-    private const string OpenAiEndpointEnvironmentVariableName = "AZURE_OPENAI_ENDPOINT";
-
     private readonly HttpClient _httpClient = httpClient;
-    private readonly IConfiguration _configuration = configuration;
+    private readonly IOpenAIConfigurationResolver _configurationResolver = configurationResolver;
     private readonly ILogger<OpenAITransportService> _logger = logger;
 
     public async Task<AIOperationResult> GenerateSummaryAsync(
@@ -37,14 +29,14 @@ internal class OpenAITransportService(
         string? promptVersion = null,
         string? fileName = null)
     {
-        var providerName = ResolveProviderName(operationName);
-        if (!string.Equals(providerName, DefaultProviderName, StringComparison.Ordinal))
+        var providerName = _configurationResolver.ResolveProviderName(operationName);
+        if (!string.Equals(providerName, "OpenAI", StringComparison.Ordinal))
         {
             _logger.LogWarning("Provider {ProviderName} is not supported by OpenAI transport.", providerName);
             return AIOperationResult.PermanentFailure(new AIProviderResult($"Unsupported provider: {providerName}"));
         }
 
-        var apiKey = ResolveApiKey(operationName);
+        var apiKey = _configurationResolver.ResolveApiKey(operationName);
         if (string.IsNullOrEmpty(apiKey))
         {
             _logger.LogWarning("Error: OpenAI API key is not configured");
@@ -64,10 +56,10 @@ internal class OpenAITransportService(
                     new { role = "system", content = resolvedSystemPrompt },
                     new { role = "user", content = content ?? string.Empty }
                 },
-                [ResolveMaxTokensParameterNameForOperation(operationName)] = maxTokens
+                [_configurationResolver.ResolveMaxTokensParameterNameForOperation(operationName)] = maxTokens
             };
 
-            var resolvedTemperature = temperature ?? ResolveConfiguredTemperature(operationName);
+            var resolvedTemperature = temperature ?? _configurationResolver.ResolveConfiguredTemperature(operationName);
             if (resolvedTemperature.HasValue)
             {
                 requestPayload["temperature"] = resolvedTemperature.Value;
@@ -76,10 +68,13 @@ internal class OpenAITransportService(
             var json = JsonSerializer.Serialize(requestPayload);
             var httpContent = new StringContent(json, Encoding.UTF8, "application/json");
 
-            _httpClient.DefaultRequestHeaders.Clear();
-            _httpClient.DefaultRequestHeaders.Add("Authorization", apiKey);
+            using var request = new HttpRequestMessage(HttpMethod.Post, _configurationResolver.ResolveApiUrl(operationName))
+            {
+                Content = httpContent
+            };
+            request.Headers.TryAddWithoutValidation("Authorization", apiKey);
 
-            var response = await _httpClient.PostAsync(ResolveApiUrl(operationName), httpContent);
+            var response = await _httpClient.SendAsync(request);
             var responseContent = await response.Content.ReadAsStringAsync();
             var metadata = TryExtractProviderMetadata(responseContent);
             var providerResponse = BuildProviderResponseFromMetadata(
@@ -129,155 +124,6 @@ internal class OpenAITransportService(
             _logger.LogError(ex, "Error generating AI summary");
             return AIOperationResult.TransientFailure(new AIProviderResult(ex.Message));
         }
-    }
-
-    private string ResolveApiKey(string? operationName = null)
-    {
-        var providerName = ResolveProviderName(operationName);
-        if (string.Equals(providerName, DefaultProviderName, StringComparison.Ordinal))
-        {
-            var injectedApiKey = _configuration[OpenAiApiKeyEnvironmentVariableName];
-            if (!string.IsNullOrWhiteSpace(injectedApiKey))
-            {
-                return injectedApiKey;
-            }
-        }
-
-        return _configuration[$"Azure:{providerName}:ApiKey"] ?? string.Empty;
-    }
-
-    private string ResolveProviderName(string? operationName = null)
-    {
-        if (!string.IsNullOrWhiteSpace(operationName))
-        {
-            var configuredProvider = _configuration[$"Azure:Operations:{operationName}:Provider"];
-            if (!string.IsNullOrWhiteSpace(configuredProvider))
-            {
-                return configuredProvider.Trim();
-            }
-        }
-
-        var defaultProvider = _configuration["Azure:Operations:Defaults:Provider"];
-        return string.IsNullOrWhiteSpace(defaultProvider) ? DefaultProviderName : defaultProvider.Trim();
-    }
-
-    private string ResolveMaxTokensParameterNameForOperation(string? operationName = null)
-    {
-        var providerName = ResolveProviderName(operationName);
-        var profileName = ResolveProfileName(operationName);
-        var profileParameterName = ResolveProfileSetting(providerName, profileName, "MaxTokensParameter");
-        return ResolveMaxTokensParameterName(profileParameterName);
-    }
-
-    private double? ResolveConfiguredTemperature(string? operationName = null)
-    {
-        var providerName = ResolveProviderName(operationName);
-        var profileName = ResolveProfileName(operationName);
-        var profileTemperature = ResolveProfileSetting(providerName, profileName, "Temperature");
-        if (profileTemperature != null
-            && double.TryParse(profileTemperature, NumberStyles.Float, CultureInfo.InvariantCulture, out var parsedTemperature))
-        {
-            return parsedTemperature;
-        }
-
-        return null;
-    }
-
-    private string ResolveApiUrl(string? operationName)
-    {
-        var providerName = ResolveProviderName(operationName);
-        var profileName = ResolveProfileName(operationName);
-        var profileApiUrl = ResolveProfileSetting(providerName, profileName, "ApiUrl");
-        var injectedEndpoint = ResolveInjectedEndpoint(providerName);
-        var legacyOpenAiApiUrl = _configuration["Azure:OpenAI:ApiUrl"];
-
-        if (!string.IsNullOrWhiteSpace(injectedEndpoint) && !string.IsNullOrWhiteSpace(profileApiUrl))
-        {
-            return CombineEndpointAndPath(injectedEndpoint, profileApiUrl);
-        }
-
-        if (!string.IsNullOrWhiteSpace(profileApiUrl))
-        {
-            return profileApiUrl;
-        }
-
-        if (!string.IsNullOrWhiteSpace(legacyOpenAiApiUrl))
-        {
-            return legacyOpenAiApiUrl;
-        }
-
-        throw new InvalidOperationException($"AI API URL is not configured for provider '{providerName}'.");
-    }
-
-    private string? ResolveInjectedEndpoint(string providerName)
-    {
-        if (!string.Equals(providerName, DefaultProviderName, StringComparison.Ordinal))
-        {
-            return _configuration[$"Azure:{providerName}:Endpoint"];
-        }
-
-        var injectedEndpoint = _configuration[OpenAiEndpointEnvironmentVariableName];
-        if (!string.IsNullOrWhiteSpace(injectedEndpoint))
-        {
-            return injectedEndpoint;
-        }
-
-        return _configuration["Azure:OpenAI:Endpoint"];
-    }
-
-    private string? ResolveProfileName(string? operationName)
-    {
-        if (!string.IsNullOrWhiteSpace(operationName))
-        {
-            var operationProfile = _configuration[$"Azure:Operations:{operationName}:Profile"];
-            if (!string.IsNullOrWhiteSpace(operationProfile))
-            {
-                return operationProfile.Trim();
-            }
-        }
-
-        var defaultProfile = _configuration["Azure:Operations:Defaults:Profile"];
-        return string.IsNullOrWhiteSpace(defaultProfile) ? null : defaultProfile.Trim();
-    }
-
-    private string? ResolveProfileSetting(string providerName, string? profileName, string settingName)
-    {
-        if (string.IsNullOrWhiteSpace(profileName))
-        {
-            return null;
-        }
-
-        var profileSetting = _configuration[$"Azure:{providerName}:Profiles:{profileName}:{settingName}"];
-        return string.IsNullOrWhiteSpace(profileSetting) ? null : profileSetting;
-    }
-
-    private static string ResolveMaxTokensParameterName(string? configuredParameterName)
-    {
-        if (string.Equals(configuredParameterName, LegacyMaxTokensParameterName, StringComparison.Ordinal))
-        {
-            return LegacyMaxTokensParameterName;
-        }
-
-        return DefaultMaxTokensParameterName;
-    }
-
-    private static string CombineEndpointAndPath(string endpoint, string profilePath)
-    {
-        const char UrlPathSeparator = '/';
-
-        if (Uri.TryCreate(profilePath, UriKind.Absolute, out var absoluteUri))
-        {
-            return absoluteUri.ToString();
-        }
-
-        var trimmedEndpoint = endpoint.Trim().TrimEnd(UrlPathSeparator);
-        var trimmedPath = profilePath.Trim();
-        if (!trimmedPath.StartsWith(UrlPathSeparator))
-        {
-            trimmedPath = string.Concat(UrlPathSeparator, trimmedPath);
-        }
-
-        return trimmedEndpoint + trimmedPath;
     }
 
     private static AIOperationResult MapFailureOutcome(HttpStatusCode statusCode, AIProviderResult response)
